moved color-picker-launcher to color-scheme folder
