+++ src/app/modules/color-scheme/color-picker-launcher/color-picker-launcher.component.ts
