+++ src/app/modules/image-viewer/image-viewer-layers/image-viewer-layers.module.ts
@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 import { MatCheckboxModule } from '@angular/material/checkbox';
 
 import { ImageViewerLayersComponent } from './image-viewer-layers.component';
-import { ColorPickerLauncherModule } from '../color-picker-launcher/color-picker-launcher.module';
+import { ColorPickerLauncherModule } from '../../color-scheme/color-picker-launcher/color-picker-launcher.module';
 
 
 @NgModule({
